Create a shared SshSessions struct.
