+++ rust/src/nasl/builtin/ssh/libssh/mod.rs
@@ -2,8 +2,6 @@
 //
 // SPDX-License-Identifier: GPL-2.0-or-later WITH x11vnc-openssl-exception
 
-// TODO make error handling less redundant
-// TODO clean up and maybe split as 2000 lines is a bit much
 //! Defines NASL ssh and sftp functions
 
 mod channel;
@@ -11,9 +9,8 @@ mod error;
 mod session;
 pub mod sessions;
 
-pub use self::sessions::Ssh;
-
 pub type SessionId = i32;
 pub type Socket = std::os::raw::c_int;
 pub use error::SshError;
 pub use libssh_rs::AuthMethods;
+pub use session::SshSession;
